pyproject.toml
[tool.poetry]
name = "xhom"
version = "0.1.0"
description = "Our Smart Home Central Control Dashboard is designed using the Flask web framework as its core with Python. It utilizes an application entry point named `app.py` that sets up Flask and incorporates routes from `routes.py` for mapping URL paths such as `/dashboard` for the main control panel access and `/devices/{id}/control` for specific device operations. The `dashboard_controller.py` is responsible for retrieving information about all connected devices from `device_model.py`, which interacts with MongoDB through an ORM-like mechanism allowing for flexible database interactions. Device commands and operations are managed via `device_control.py`, communicating with smart devices using a universal API adapter. The device and user settings, including automated routine scripts, are maintained in MongoDB through `device_model.py`, facilitating dynamic schema modifications. The frontend is powered by Vue.js, offering a reactive and component-based structure, enriched with Vuex for state management and Vue Router for single-page application (SPA) features. AJAX calls are implemented within Vue components to provide a seamless user experience with real-time feedback. Additionally, voice command features are incorporated through `voice_handlers.py`, leveraging voice assistant APIs like those from Google Assistant or Amazon Alexa to convert spoken commands into dashboard actions. This configuration not only streamlines home automation management for users but also lays down a strong, scalable foundation for further advancements and integrations within the Internet of Things (IoT) ecosystem."
authors = ["AutoGPT <info@agpt.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11"
bcrypt = "^3.2.0"
datetime = "*"
fastapi = "*"
passlib = {version = "*", extras = ["bcrypt"]}
prisma = "*"
pydantic = "*"
python-jose = {version = "*", extras = ["cryptography"]}
uvicorn = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
